rename ImageManage.ts
