+++ src/image_manager.ts
