+++ src/view/canvas_components/canvas.ts
@@ -1,6 +1,6 @@
 import {fabric} from 'fabric';
 
-import {GetImage} from '../../ImageManager';
+import {GetImage} from '../../image_manager';
 import {Point} from '../../utils/index';
 
 import {Accessory} from './accessory';
